MAKE Quiz component for refactoring
App.js에서 Rougnt 관련 코드를 작성하기 위해 App.js에 있던 기존의 코드를 모두 Quiz 컴포넌트로 옮겼다.
